http: add expect/continue
When sending POST requests, we may encounter a 401 unauthorized
response.  It's - at best - disappointing to send the POST body
to the server and then get a response that requires us to replay
it; but it's - at worst - impossible to actually replay this because
we were streaming it to the server and can't rewind that buffer
easily to resend it.

Expect/continue allows us to send only the headers to the server,
keeping the POST body buffered.  This lets us negotiate authentication
without advancing our stream.  Once authentication completes (we get a
"100 Continue" response from the server) then we can send the POST data.